+++ inc/ticket.class.php
@@ -29,6 +29,7 @@
  */
 
 use Glpi\Application\View\TemplateRenderer;
+use Glpi\Toolbox\Sanitizer;
 
 if (!defined('GLPI_ROOT')) {
     die("Sorry. You can't access directly to this file");
@@ -38,7 +39,7 @@ class PluginEscaladeTicket
 {
     public static function pre_item_update(CommonDBTM $item)
     {
-        if (isset($input['_itil_assign'])) {
+        if (isset($item->input['_itil_assign'])) {
             $item->input['_do_not_compute_status'] = true;
         }
         $config = $_SESSION['plugins']['escalade']['config'];
@@ -108,6 +109,8 @@ class PluginEscaladeTicket
                     }
                 }
             }
+
+            self::removeAssignUsers($item);
             return $item;
         }
     }
@@ -418,7 +421,7 @@ class PluginEscaladeTicket
      */
     public static function assignUserGroup(Ticket $ticket)
     {
-        if (!is_array($ticket->input) || !count($ticket->input)) {
+        if (!count($ticket->input)) {
             // Already cancel by another plugin
             return false;
         }
@@ -485,6 +488,8 @@ class PluginEscaladeTicket
     public static function climb_group($tickets_id, $groups_id)
     {
         //don't add group if already exist for this ticket
+        $group = new Group();
+        $group->getFromDB($groups_id);
         $group_ticket = new Group_Ticket();
         $condition = [
             'tickets_id' => $tickets_id,
@@ -492,25 +497,38 @@ class PluginEscaladeTicket
             'type'       => CommonITILActor::ASSIGN
         ];
         if (!$group_ticket->find($condition)) {
-            $ticket = new Ticket();
-            $ticket->getFromDB($tickets_id);
+            $ticket_group = new Group_Ticket();
+            if (
+                $ticket_group->add(
+                    [
+                        'tickets_id'                    => $tickets_id,
+                        'groups_id'                     => $groups_id,
+                        'type'                          => CommonITILActor::ASSIGN,
+                        '_disablenotif'                 => true,
+                        '_plugin_escalade_no_history'   => true,
+                    ]
+                )
+            ) {
+                if ($_SESSION['plugins']['escalade']['config']['task_history']) {
+                    $task = new TicketTask();
+                    $task->add([
+                        'tickets_id' => $tickets_id,
+                        'is_private' => true,
+                        'state'      => Planning::INFO,
+                        // Sanitize before merging with $_POST['comment'] which is already sanitized
+                        'content'    => Sanitizer::sanitize(
+                            '<p><i>' . sprintf(__('Escalation to the group %s.', 'escalade'), Sanitizer::unsanitize($group->getName())) . '</i></p><hr />'
+                        ) . $_POST['comment']
+                    ]);
+                }
 
-            // Update the ticket with actor data in order to execute the necessary rules
-            $_form_object = [
-                '_do_not_compute_status' => true,
-            ];
-            if ($_SESSION['plugins']['escalade']['config']['ticket_last_status'] != -1) {
-                $_form_object['status'] = $_SESSION['plugins']['escalade']['config']['ticket_last_status'];
+                //notified only the last group assigned
+                $ticket = new Ticket();
+                $ticket->getFromDB($tickets_id);
+
+                $event = "assign_group";
+                NotificationEvent::raiseEvent($event, $ticket);
             }
-            $ticket_details = $_POST['ticket_details'] ?? $_GET['ticket_details'] ?? [];
-            $updates_ticket = new Ticket();
-            $updates_ticket->update($ticket_details + [
-                '_actors' => PluginEscaladeTicket::getTicketFieldsWithActors($tickets_id, $groups_id),
-                '_plugin_escalade_no_history' => false,
-                'actortype' => CommonITILActor::ASSIGN,
-                'groups_id' => $groups_id,
-                '_form_object' => $_form_object,
-            ]);
         }
 
         Html::back();
@@ -554,13 +572,13 @@ class PluginEscaladeTicket
             $_SESSION['plugins']['escalade']['config']['remove_tech'] == false
             && $_SESSION['plugins']['escalade']['config']['remove_requester'] == false
         ) {
-            return true;
+            return;
         }
         if ($type == CommonITILActor::ASSIGN && !$_SESSION['plugins']['escalade']['config']['remove_tech']) {
-            return true;
+            return;
         }
         if ($type == CommonITILActor::REQUESTER && !$_SESSION['plugins']['escalade']['config']['remove_requester']) {
-            return true;
+            return;
         }
 
         $tickets_id = $item->input['id'] ?? $item->fields['id'];
@@ -592,7 +610,6 @@ class PluginEscaladeTicket
 
             //delete user
             $ticket_user->delete(['id' => $id]);
-
             if (isset($item->input['_actors'])) {
                 foreach ($item->input['_actors'][$types[$type]] as $key => $actor) {
                     if (
@@ -617,7 +634,7 @@ class PluginEscaladeTicket
      * Update ticket status when user added.
      * Trigger also adding user groups if feature enabled
      * @param  Ticket_User $item Ticket_User object
-     * @return void
+     * @return bool
      */
     public static function item_add_user(Ticket_User $item, $type = CommonITILActor::ASSIGN)
     {
@@ -627,13 +644,13 @@ class PluginEscaladeTicket
         $ticket->getFromDB($tickets_id);
         $groups_id = [];
 
-        self::removeAssignUsers($ticket, $users_id, $type);
 
         // == Add user groups on modification ==
         //check this plugin config
         if (
             $_SESSION['plugins']['escalade']['config']['use_assign_user_group'] == 0
             || $_SESSION['plugins']['escalade']['config']['use_assign_user_group_modification'] == 0
+            || $type != CommonITILActor::ASSIGN
         ) {
             return true;
         }
@@ -664,7 +681,7 @@ class PluginEscaladeTicket
                 'type'       => CommonITILActor::ASSIGN
             ]);
             if (!empty($found)) {
-                return;
+                return false;
             }
 
             //prevent user removal
@@ -684,7 +701,7 @@ class PluginEscaladeTicket
         }
 
         //fix ticket status
-        $ticket->update([
+        return $ticket->update([
             'id'     => $tickets_id,
             'status' => CommonITILObject::ASSIGNED
         ]);
@@ -693,7 +710,9 @@ class PluginEscaladeTicket
 
     /**
      * Close linked tickets when ticket passed in parameter is closed
-     * @param  CommonDBTM $item the ticket object
+     * @param  CommonDBTM $ticket the ticket object
+     * @param  int $status
+     *
      * @return void
      */
     public static function linkedTickets(CommonDBTM $ticket, $status = CommonITILObject::SOLVED)
@@ -733,7 +752,7 @@ class PluginEscaladeTicket
         //get auto-assign mode (config in entity)
         $auto_assign_mode = Entity::getUsedConfig('auto_assign_mode', $_SESSION['glpiactive_entity']);
         if ($auto_assign_mode == Entity::CONFIG_NEVER) {
-            return true;
+            return;
         }
 
         //get category
@@ -871,8 +890,7 @@ class PluginEscaladeTicket
       SELECT null AS id, $newID as tickets_id, users_id, type, use_notification, alternative_email
       FROM glpi_tickets_users
       WHERE tickets_id = $tickets_id AND type != 2";
-        // @phpstan-ignore-next-line
-        if (!$res = $DB->query($query_users)) {
+        if (!$res = $DB->doQuery($query_users)) {
             Session::addMessageAfterRedirect(__('Error : adding actors (user)', 'escalade'), false, ERROR);
             exit;
         }
@@ -881,8 +899,7 @@ class PluginEscaladeTicket
       SELECT null AS id, $newID as tickets_id, groups_id, type
       FROM glpi_groups_tickets
       WHERE tickets_id = $tickets_id AND type != 2";
-        // @phpstan-ignore-next-line
-        if (!$res = $DB->query($query_groups)) {
+        if (!$res = $DB->doQuery($query_groups)) {
             Session::addMessageAfterRedirect(__('Error : adding actors (group)', "escalade"), false, ERROR);
             exit;
         }
@@ -892,8 +909,7 @@ class PluginEscaladeTicket
       SELECT documents_id, $newID, 'Ticket', entities_id, is_recursive, date_mod
       FROM glpi_documents_items
       WHERE items_id = $tickets_id AND itemtype = 'Ticket'";
-        // @phpstan-ignore-next-line
-        if (!$res = $DB->query($query_docs)) {
+        if (!$res = $DB->doQuery($query_docs)) {
             Session::addMessageAfterRedirect(__('Error : adding documents', 'escalade'), false, ERROR);
             exit;
         }
@@ -1069,7 +1085,7 @@ class PluginEscaladeTicket
 
         $_actors['assign'] = array_merge(
             $ticket_actors['User']['assign'] ?? [],
-            $ticket_actors['Group']['assign'] ?? [],
+            $ticket_actors['Group']['assign'],
             $ticket_actors['Supplier']['assign'] ?? []
         );
         $_actors['observer'] = array_merge(